test(exec/protocol): add focused protocol proxy tests and refactor helpers
- Removed legacy proxy_helpers.rs in tests and replaced with a modular proxy_helpers
  subdirectory structure: error_mapping, forwarding, routing, and a mod.rs for shared helpers.
- Introduced detailed tests covering stdin forwarding, stdout/stderr routing, and error mappings.
- Refactored protocol exec test code to utilize new RecordingWriter types with failure modes.
- Improved test clarity by removing unused scenario writers and consolidating common test setup.
- Added a timeout constant and more explicit error classification for stdin forwarding in protocol.rs.
